fix(ux): /build crash + error boundary escalation chain to GitHub
Two fixes for the Arcamanus customer-0 install:

1. FleetDensityBar: `for (const r of taskResults ?? [])` crashes when a
   featureBuild has taskResults stored as `{}` (empty JSON object, not null
   and not an array). `{} ?? []` returns `{}` -- not iterable. Guard with
   `Array.isArray` so non-array JSON values are treated as empty.

2. Error boundary UX chain: the crash error page auto-files a PIR and lets
   the user add context, but had no "Report to the project team" button --
   leaving a non-technical operator with no path from crash -> GitHub BI.
   - Extract UpstreamEscalation from FeedbackForm into its own component
     so it can be shared by both FeedbackForm and the crash error boundary.
   - Show UpstreamEscalation in error.tsx: immediately after the
     auto-report is filed (autoReportId), and again after the user submits
     their description. Full chain: crash -> PIR -> "Report to the project
     team" -> consent -> GitHub issue.
   - Collapse the "Developer diagnostic / Copy prompt" section by default
     behind a disclosure triangle so the primary CTA for a non-technical
     user is the description field + escalation button.
